Add files via upload
Matlab app As proof of concept, here is a simple Matlab GUI app demonstration. Using earlier released code Matlab App designer was used to create a gui panel to enter callsigns and grid square locators and then transmit a cq call on FT8 using these. A flowgraph in GNU Radio Companion is used to take audio produced in Matlab and modulate an RF carrier producing an USB signal that is sent to the HackRF One. Transmissions at low power and with a short wire antenna are monitored using an rtl-sdr dongle via SDRSharp, VB Audiocable and WSJTX.